+++ example/custom_claims.py
@@ -104,7 +104,6 @@ def login():
     """
     Logs a user in by parsing a POST request containing user credentials and
     issuing a JWT token.
-
     .. example::
        $ curl http://localhost:5000/login -X POST \
          -d '{"username":"Walter","password":"calmerthanyouare"}'
@@ -130,7 +129,6 @@ def protected():
     """
     A protected endpoint. The auth_required decorator will require a header
     containing a valid JWT
-
     .. example::
        $ curl http://localhost:5000/protected -X GET \
          -H "Authorization: Bearer <your_token>"