+++ example/refresh.py
@@ -90,7 +90,6 @@ def login():
     """
     Logs a user in by parsing a POST request containing user credentials and
     issuing a JWT token.
-
     .. example::
        $ curl http://localhost:5000/login -X POST \
          -d '{"username":"Walter","password":"calmerthanyouare"}'
@@ -108,7 +107,6 @@ def refresh():
     """
     Refreshes an existing JWT by creating a new one that is a copy of the old
     except that it has a refrehsed access expiration.
-
     .. example::
        $ curl http://localhost:5000/refresh -X GET \
          -H "Authorization: Bearer <your_token>"
@@ -125,7 +123,6 @@ def protected():
     """
     A protected endpoint. The auth_required decorator will require a header
     containing a valid JWT
-
     .. example::
        $ curl http://localhost:5000/protected -X GET \
          -H "Authorization: Bearer <your_token>"
@@ -141,7 +138,6 @@ def protected():
 def disable_user():
     """
     Disables a user in the data store
-
     .. example::
         $ curl http://localhost:5000/disable_user -X POST \
           -H "Authorization: Bearer <your_token>" \