+++ example/static/custom.js
@@ -1,7 +1,0 @@
-function getProtected() {
-  makeRequest(
-    'GET',
-    'protected',
-    addTokenHeader(),
-  );
-};