+++ flask_praetorian/base.py
@@ -1,4 +1,6 @@
+import datetime
 import flask
+import jinja2
 import jwt
 import pendulum
 import re
@@ -6,13 +8,12 @@ import textwrap
 import uuid
 import warnings
 
-from jinja2 import Template
-from flask import url_for
 from flask_mail import Message
 
 from passlib.context import CryptContext
 
-from flask_praetorian.utilities import deprecated
+from flask_praetorian.utilities import deprecated, duration_from_string
+
 from flask_praetorian.exceptions import (
     AuthenticationError,
     BlacklistedError,
@@ -20,13 +21,16 @@ from flask_praetorian.exceptions import (
     EarlyRefreshError,
     ExpiredAccessError,
     ExpiredRefreshError,
+    InvalidRegistrationToken,
     InvalidTokenHeader,
     InvalidUserError,
+    LegacyScheme,
     MissingClaimError,
     MissingTokenHeader,
     MissingUserError,
+    MisusedRegistrationToken,
+    ConfigurationError,
     PraetorianError,
-    LegacyScheme,
 )
 
 from flask_praetorian.constants import (
@@ -37,15 +41,15 @@ from flask_praetorian.constants import (
     DEFAULT_JWT_HEADER_TYPE,
     DEFAULT_JWT_REFRESH_LIFESPAN,
     DEFAULT_USER_CLASS_VALIDATION_METHOD,
-    DEFAULT_EMAIL_TEMPLATE,
-    DEFAULT_CONFIRMATION_ENDPOINT,
-    DEFAULT_CONFIRMATION_SENDER,
+    DEFAULT_CONFIRMATION_TEMPLATE,
     DEFAULT_CONFIRMATION_SUBJECT,
     DEFAULT_HASH_SCHEME,
     DEFAULT_HASH_ALLOWED_SCHEMES,
     DEFAULT_HASH_AUTOUPDATE,
     DEFAULT_HASH_AUTOTEST,
     DEFAULT_HASH_DEPRECATED_SCHEMES,
+    IS_REGISTRATION_TOKEN_CLAIM,
+    REFRESH_EXPIRATION_CLAIM,
     RESERVED_CLAIMS,
     VITAM_AETERNUM,
     AccessType,
@@ -65,7 +69,7 @@ class Praetorian:
         self.salt = None
 
         if app is not None and user_class is not None:
-            self.app = self.init_app(app, user_class, is_blacklisted)
+            self.init_app(app, user_class, is_blacklisted)
 
     def init_app(self, app, user_class, is_blacklisted=None):
         """
@@ -125,14 +129,14 @@ class Praetorian:
             'JWT_ALGORITHM',
             DEFAULT_JWT_ALGORITHM,
         )
-        self.access_lifespan = pendulum.Duration(**app.config.get(
+        self.access_lifespan = app.config.get(
             'JWT_ACCESS_LIFESPAN',
             DEFAULT_JWT_ACCESS_LIFESPAN,
-        ))
-        self.refresh_lifespan = pendulum.Duration(**app.config.get(
+        )
+        self.refresh_lifespan = app.config.get(
             'JWT_REFRESH_LIFESPAN',
             DEFAULT_JWT_REFRESH_LIFESPAN,
-        ))
+        )
         self.header_name = app.config.get(
             'JWT_HEADER_NAME',
             DEFAULT_JWT_HEADER_NAME,
@@ -145,21 +149,38 @@ class Praetorian:
             'USER_CLASS_VALIDATION_METHOD',
             DEFAULT_USER_CLASS_VALIDATION_METHOD,
         )
-        self.email_template = app.config.get(
-            'PRAETORIAN_EMAIL_TEMPLATE',
-            DEFAULT_EMAIL_TEMPLATE,
+
+        self.confirmation_template = app.config.get(
+            'PRAETORIAN_CONFIRMATION_TEMPLATE',
+            DEFAULT_CONFIRMATION_TEMPLATE,
         )
-        self.confirmation_endpoint = app.config.get(
-            'PRAETORIAN_CONFIRMATION_ENDPOINT',
-            DEFAULT_CONFIRMATION_ENDPOINT,
+        self.confirmation_uri = app.config.get(
+            'PRAETORIAN_CONFIRMATION_URI',
         )
         self.confirmation_sender = app.config.get(
             'PRAETORIAN_CONFIRMATION_SENDER',
-            DEFAULT_CONFIRMATION_SENDER,
         )
         self.confirmation_subject = app.config.get(
             'PRAETORIAN_CONFIRMATION_SUBJECT',
             DEFAULT_CONFIRMATION_SUBJECT,
+        )
+
+        if isinstance(self.access_lifespan, dict):
+            self.access_lifespan = pendulum.duration(**self.access_lifespan)
+        elif isinstance(self.access_lifespan, str):
+            self.access_lifespan = duration_from_string(self.access_lifespan)
+        ConfigurationError.require_condition(
+            isinstance(self.access_lifespan, datetime.timedelta),
+            "access lifespan was not configured",
+        )
+
+        if isinstance(self.refresh_lifespan, dict):
+            self.refresh_lifespan = pendulum.duration(**self.refresh_lifespan)
+        if isinstance(self.refresh_lifespan, str):
+            self.refresh_lifespan = duration_from_string(self.refresh_lifespan)
+        ConfigurationError.require_condition(
+            isinstance(self.refresh_lifespan, datetime.timedelta),
+            "refresh lifespan was not configured",
         )
 
         if not app.config.get('DISABLE_PRAETORIAN_ERROR_HANDLER'):
@@ -298,7 +319,7 @@ class Praetorian:
     def encode_jwt_token(
             self, user,
             override_access_lifespan=None, override_refresh_lifespan=None,
-            bypass_user_check=False,
+            bypass_user_check=False, is_registration_token=False,
             **custom_claims
     ):
         """
@@ -317,6 +338,9 @@ class Praetorian:
         :param: bypass_user_check:         Override checking the user for
                                            being real/active.  Used for
                                            registration token generation.
+        :param: is_registration_token:     Indicates that the token will be
+                                           used only for email-based
+                                           registration
         :param: custom_claims:             Additional claims that should
                                            be packed in the payload. Note that
                                            any claims supplied here must be
@@ -346,15 +370,20 @@ class Praetorian:
             refresh_expiration,
         )
 
-        payload_parts = dict(
-            iat=moment.int_timestamp,
-            exp=access_expiration,
-            rf_exp=refresh_expiration,
-            jti=str(uuid.uuid4()),
-            id=user.identity,
-            rls=','.join(user.rolenames),
-            **custom_claims
+        payload_parts = {
+            'iat': moment.int_timestamp,
+            'exp': access_expiration,
+            'jti': str(uuid.uuid4()),
+            'id': user.identity,
+            'rls': ','.join(user.rolenames),
+            REFRESH_EXPIRATION_CLAIM: refresh_expiration,
+        }
+        if is_registration_token:
+            payload_parts[IS_REGISTRATION_TOKEN_CLAIM] = True
+        flask.current_app.logger.debug(
+            "Attaching custom claims: {}".format(custom_claims),
         )
+        payload_parts.update(custom_claims)
         return jwt.encode(
             payload_parts, self.encode_key, self.encode_algorithm,
         ).decode('utf-8')
@@ -393,16 +422,7 @@ class Praetorian:
                                            exceed the refresh lifespan
         """
         moment = pendulum.now('UTC')
-        # Note: we disable exp verification because we do custom checks here
-        with InvalidTokenHeader.handle_errors('failed to decode JWT token'):
-            data = jwt.decode(
-                token,
-                self.encode_key,
-                algorithms=self.allowed_algorithms,
-                options={'verify_exp': False},
-            )
-
-        self._validate_jwt_data(data, access_type=AccessType.refresh)
+        data = self.extract_jwt_token(token, access_type=AccessType.refresh)
 
         user = self.user_class.identify(data['id'])
         self._check_user(user)
@@ -411,7 +431,7 @@ class Praetorian:
             access_lifespan = self.access_lifespan
         else:
             access_lifespan = override_access_lifespan
-        refresh_expiration = data['rf_exp']
+        refresh_expiration = data[REFRESH_EXPIRATION_CLAIM]
         access_expiration = min(
             (moment + access_lifespan).int_timestamp,
             refresh_expiration,
@@ -420,20 +440,20 @@ class Praetorian:
         custom_claims = {
             k: v for (k, v) in data.items() if k not in RESERVED_CLAIMS
         }
-        payload_parts = dict(
-            iat=moment.int_timestamp,
-            exp=access_expiration,
-            rf_exp=refresh_expiration,
-            jti=data['jti'],
-            id=data['id'],
-            rls=','.join(user.rolenames),
-            **custom_claims
-        )
+        payload_parts = {
+            'iat': moment.int_timestamp,
+            'exp': access_expiration,
+            'jti': data['jti'],
+            'id': data['id'],
+            'rls': ','.join(user.rolenames),
+            REFRESH_EXPIRATION_CLAIM: refresh_expiration,
+        }
+        payload_parts.update(custom_claims)
         return jwt.encode(
             payload_parts, self.encode_key, self.encode_algorithm,
         ).decode('utf-8')
 
-    def extract_jwt_token(self, token):
+    def extract_jwt_token(self, token, access_type=AccessType.access):
         """
         Extracts a data dictionary from a jwt token
         """
@@ -445,7 +465,7 @@ class Praetorian:
                 algorithms=self.allowed_algorithms,
                 options={'verify_exp': False},
             )
-        self._validate_jwt_data(data, access_type=AccessType.access)
+        self._validate_jwt_data(data, access_type=access_type)
         return data
 
     def _validate_jwt_data(self, data, access_type):
@@ -469,23 +489,40 @@ class Praetorian:
             'Token is missing exp claim',
         )
         MissingClaimError.require_condition(
-            'rf_exp' in data,
-            'Token is missing rf_exp claim',
+            REFRESH_EXPIRATION_CLAIM in data,
+            'Token is missing {} claim'.format(REFRESH_EXPIRATION_CLAIM),
         )
         moment = pendulum.now('UTC').int_timestamp
         if access_type == AccessType.access:
+            MisusedRegistrationToken.require_condition(
+                IS_REGISTRATION_TOKEN_CLAIM not in data,
+                "registration token used for access"
+            )
             ExpiredAccessError.require_condition(
                 moment <= data['exp'],
                 'access permission has expired',
             )
         elif access_type == AccessType.refresh:
+            MisusedRegistrationToken.require_condition(
+                IS_REGISTRATION_TOKEN_CLAIM not in data,
+                "registration token used for refresh"
+            )
             EarlyRefreshError.require_condition(
                 moment > data['exp'],
                 'access permission for token has not expired. may not refresh',
             )
             ExpiredRefreshError.require_condition(
-                moment <= data['rf_exp'],
+                moment <= data[REFRESH_EXPIRATION_CLAIM],
                 'refresh permission for token has expired',
+            )
+        elif access_type == AccessType.register:
+            ExpiredAccessError.require_condition(
+                moment <= data['exp'],
+                'register permission has expired',
+            )
+            InvalidRegistrationToken.require_condition(
+                IS_REGISTRATION_TOKEN_CLAIM in data,
+                "invalid registration token used for verification"
             )
 
     def _unpack_header(self, headers):
@@ -545,8 +582,9 @@ class Praetorian:
         return {self.header_name: self.header_type + ' ' + token}
 
     def send_registration_email(
-        self, user=None, template=None,
-        subject=None, override_access_lifepsan=None
+        self, email, user=None, template=None,
+        confirmation_sender=None, confirmation_uri=None,
+        subject=None, override_access_lifespan=None
     ):
         """
         Sends a registration email to a new user, containing a time expiring
@@ -558,63 +596,102 @@ class Praetorian:
             `result` from mail send.
         :param: user:                     The user object to tie claim to
                                           (username, id, email, etc)
-        :param: subject:                  The registration email subject
-                                          override.
-        :param: override_access_lifepsan: Anything used in `encode_jwt_token`,
-                                          but all we really care about is
-                                          `override_access_token`.  This
-                                          overrides the instance's defined
-                                          lifespan, which is used as the
-                                          default.  The token will no longer
-                                          be valid after this time and
-                                          be rejected.
-        :param: template:                 Template file location for email.
+        :param: template:                 HTML Template for confirmation email.
                                           If not provided, a stock entry is
-                                          used.
+                                          used
+        :param: confirmation_sender:      The sender that shoudl be attached
+                                          to the confirmation email. Overrides
+                                          the PRAETORIAN_CONFIRMRATION_SENDER
+                                          config setting
+        :param: confirmation_uri:         The uri that should be visited to
+                                          complete email registration. Should
+                                          usually be a uri to a frontend or
+                                          external service that calls a
+                                          'finalize' method in the api to
+                                          complete registration. Will override
+                                          the PRAETORIAN_CONFIRMATION_URI
+                                          config setting
+        :param: subject:                  The registration email subject.
+                                          Will override the
+                                          PRAETORIAN_CONFIRMATION_SUBJECT
+                                          config setting.
+        :param: override_access_lifespan: Overrides the JWT_ACCESS_LIFESPAN
+                                          to set an access lifespan for the
+                                          registration token.
         """
+        if subject is None:
+            subject = self.confirmation_subject
+
+        if confirmation_uri is None:
+            confirmation_uri = self.confirmation_uri
+
         notification = {
                 'result': None,
                 'message': None,
-                'email': user.email,
+                'user': str(user),
+                'email': email,
                 'token': None,
-                'confirmation_uri': None,
-                'subject': subject if subject else self.confirmation_subject,
+                'subject': subject,
+                'confirmation_uri': confirmation_uri,
         }
 
+        sender = confirmation_sender or self.confirmation_sender
+        PraetorianError.require_condition(
+            sender,
+            "A confirmation sender is required to send confirmation email",
+        )
+
+        if template is None:
+            with open(self.confirmation_template) as fh:
+                template = fh.read()
+
         with PraetorianError.handle_errors('fail sending confirmation email'):
+            app = flask.current_app
+            app.logger.debug("NOTIFICATION: {}".format(notification))
+            app.logger.debug(
+                "Generating registration token with lifespan: {}".format(
+                    override_access_lifespan
+                )
+            )
             notification['token'] = self.encode_jwt_token(
                 user,
-                override_access_lifepsan,
-                bypass_user_check=True
+                override_access_lifespan=override_access_lifespan,
+                bypass_user_check=True, is_registration_token=True,
             )
-            try:
-                """ attempt to find the URI for registration confirmation """
-                _confirmation_uri = url_for(self.confirmation_endpoint,
-                                            _external=True)
-                notification['confirmation_uri'] = '/'.join(
-                        [_confirmation_uri, notification['token']]
-                )
-            except Exception:
-                """ if fails, lets set None """
-                notification['confirmation_uri'] = None
-
-            with open(self.email_template) as _template:
-                tmpl = Template(_template.read())
-            notification['message'] = tmpl.render(notification).strip()
+            jinja_tmpl = jinja2.Template(template)
+            notification['message'] = jinja_tmpl.render(notification).strip()
 
             msg = Message(
                     html=notification['message'],
-                    sender=self.confirmation_sender,
+                    sender=sender,
                     subject=notification['subject'],
                     recipients=[notification['email']]
             )
 
-            notification['result'] = self.app.extensions['mail'].send(msg)
+            app.logger.debug("Sending verification email to {}".format(email))
+            notification['result'] = app.extensions['mail'].send(msg)
 
         return notification
 
-    def validate_confirmation(self, token):
-        return self.extract_jwt_token(token)
+    def get_user_from_registration_token(self, token):
+        """
+        Gets a user based on the registration token that is supplied. Verifies
+        that the token is a regisration token and that the user can be properly
+        retrieved
+        """
+        data = self.extract_jwt_token(token, access_type=AccessType.register)
+        flask.current_app.logger.debug("DATA: {}".format(data))
+        user_id = data.get('id')
+        PraetorianError.require_condition(
+            user_id is not None,
+            "Could not fetch an id from the registration token",
+        )
+        user = self.user_class.identify(user_id)
+        PraetorianError.require_condition(
+            user is not None,
+            "Could not identify the user from the registration token",
+        )
+        return user
 
     def hash_password(self, raw_password):
         """